requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
Django==3.0.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
mysqlclient==1.4.6
Pillow==7.1.2
pytz==2020.1
soupsieve==2.0
sqlparse==0.3.1
xlwt==1.3.0
